Removed unnecessary floating point conversions.

+++ src/ui/gtk/filter.c
@@ -502,7 +502,7 @@ filter_gui_filter_add(filter_t *f, GList *ruleset)
         } else {
             gm_snprintf(buf[2], sizeof buf[2], "%d/%d (%d%%)",
                 f->match_count, count,
-				(gint) (f->match_count * 100.0 / count));
+				f->match_count * 100 / count);
         }
         titles[2] = buf[2];
     } else {
@@ -559,7 +559,7 @@ filter_gui_filter_add(filter_t *f, GList *ruleset)
         } else {
             gm_snprintf(buf, sizeof buf, "%d/%d (%d%%)",
                 f->match_count, count,
-				(gint) (f->match_count * 100.0 / count));
+				f->match_count * 100 / count);
         }
     } else {
 		g_strlcpy(buf, "...", sizeof buf);
@@ -918,8 +918,8 @@ filter_get_filter_stats(const filter_t *filter)
 				gm_snprintf(buf, sizeof buf, "%d", filter->match_count);
 			} else {
 				gm_snprintf(buf, sizeof buf, "%d/%d (%d%%)",
-						filter->match_count, n,
-						(gint) (filter->match_count * 100.0 / n));
+					filter->match_count, n,
+					filter->match_count * 100 / n);
 			}
 			title = buf;
 		} else {
@@ -1018,8 +1018,8 @@ filter_get_rule_stats(const rule_t *rule)
 		n = rule->match_count + rule->fail_count;
 		if (n != 0) {
 			gm_snprintf(buf, sizeof buf, "%d/%d (%d%%)",
-					rule->match_count, n,
-					(gint) (rule->match_count * 100.0 / n));
+				rule->match_count, n,
+				rule->match_count * 100 / n);
 			title = buf;
 		} else {
 			title = _("none yet");
